Fix abs warnings

+++ src/a_action.c
@@ -187,8 +187,9 @@ void A_PotteryCheck(mobj_t *actor)
 	if (!netgame) {
 		pmo = players[consoleplayer].mo;
 		if (P_CheckSight(actor, pmo) &&
-		    (abs(R_PointToAngle2(pmo->x, pmo->y, actor->x, actor->y) -
-			 pmo->angle) <=
+		    (abs((int)R_PointToAngle2(pmo->x, pmo->y, actor->x,
+					      actor->y) -
+			 (int)pmo->angle) <=
 		     ANGLE_45)) { // Previous state (pottery bit waiting state)
 			P_SetMobjState(actor, actor->state - &states[0] - 1);
 		} else {
@@ -201,9 +202,9 @@ void A_PotteryCheck(mobj_t *actor)
 			}
 			pmo = players[i].mo;
 			if (P_CheckSight(actor, pmo) &&
-			    (abs(R_PointToAngle2(pmo->x, pmo->y, actor->x,
-						 actor->y) -
-				 pmo->angle) <=
+			    (abs((int)R_PointToAngle2(pmo->x, pmo->y, actor->x,
+						      actor->y) -
+				 (int)pmo->angle) <=
 			     ANGLE_45)) { // Previous state (pottery bit waiting state)
 				P_SetMobjState(actor,
 					       actor->state - &states[0] - 1);
